+++ src/betterKickAPI/eventsub/__init__.py
@@ -1,4 +1,4 @@
 # noqa: N999
-from betterKickAPI.eventsub.webhook import KickWebhook
+from betterKickAPI.eventsub.webhook import KickWebhook, SSLOptions, VerificationHeaders
 
-__all__ = ["KickWebhook"]
+__all__ = ["KickWebhook", "SSLOptions", "VerificationHeaders"]